Implement proper handling of here documents
This changes the approach from being a single search-and-replace
which rewrites the code slightly by outdenting all here documents
as documented in Text::Balanced to an approach that splits up
the lines, replaces the placeholders back with their values and
respects here document strings and treats them like they should
be treated.

This means that line numbers in error messages match up the
source code even if here documents are used.

This adresses RT #119932